requirements.txt
aerich==0.9.2
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncclick==8.3.0.3
asyncpg==0.30.0
click==8.3.1
dictdiffer==0.9.0
exceptiongroup==1.3.1
fastapi==0.121.3
h11==0.16.0
idna==3.11
iso8601==2.1.0
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pypika-tortoise==0.6.2
python-dotenv==1.2.1
pytz==2025.2
sniffio==1.3.1
starlette==0.50.0
tomlkit==0.13.3
tortoise-orm==0.25.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
